argon/requirements.txt
--only-binary=:all:
aerich>=0.7.1
aiocache[msgpack]>=0.12.1
aiohttp-asgi>=0.5.2
aiohttp>=3.10.0
aiosqlite>=0.17.0
anyio>=3.7.0
asyncpg>=0.29.0
bidict>=0.22.1
braceexpand>=0.1.7
colorama>=0.4.6
dateparser>=1.1.8
dictdiffer>=0.9.0
discord.py>=2.3.0
fastapi>=0.98.0
humanize>=4.6.0
imagehash>=4.3.1
import-expression>=1.1.4
iso8601>=1.1.0
jinja2>=3.1.2
jishaku>=2.5.1
lru-dict>=1.3.0
msgpack>=1.0.5
numpy>=1.26.0
parsedatetime>=2.6
pillow>=10.4.0
prettytable>=3.8.0
psutil>=5.9.5
pydantic>=1.10.9
pynacl>=1.5.0
pypika-tortoise>=0.1.6
pypika>=0.48.9
python-dateutil>=2.8.2
python-engineio>=4.4.1
python-multipart>=0.0.6
python-socketio>=5.8.0
pytz>=2023.3
pywavelets>=1.5.0
regex>=2023.6.3
scipy>=1.14.0
starlette>=0.27.0
tomlkit>=0.11.8
tortoise-orm>=0.19.3
typing-extensions>=4.6.3
tzdata>=2023.3
tzlocal>=5.0.1
ujson>=5.10.0
wcwidth>=0.2.6
yarl>=1.9.4
cffi>=1.17.0
aiohttp>=3.9.5
asyncpg>=0.29.0
cryptography>=42.0.5
